Valgrind fixes

+++ src/server/game/Entities/Player/Player.h
@@ -2985,7 +2985,7 @@ class Player : public Unit, public GridObject<Player>
 
         bool IsAlwaysDetectableFor(WorldObject const* seer) const;
 
-        uint8 m_grantableLevels;
+        uint32 m_grantableLevels;
 
         uint32 m_DelayedOperations;
